requirements.txt
absl-py
astor
attrs
backcall
bleach
certifi
chardet
click
colorama
contextlib2
cycler
cython
decorator
defusedxml
entrypoints
flask
flask-cors
gast
google-pasta
grpcio
h5py
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
jinja2
jsonschema
keras-applications
keras-preprocessing
kiwisolver
lxml
markdown
markupsafe
matplotlib
mistune
more-itertools
nbconvert
nbformat
notebook
numpy>=1.26,<2
opencv-python-headless>=4.10,<5
pandocfilters
parso
pickleshare
pillow>=10,<11
prometheus-client
prompt-toolkit
protobuf>=4,<5
pygments
pyparsing
pyrsistent
python-dateutil
pyzmq
qtconsole
requests
send2trash
six
tensorboard
tensorflow-cpu==2.16.1 ; sys_platform != "darwin"
tensorflow-macos==2.16.1 ; sys_platform == "darwin"
tensorflow-metal==1.2.0 ; sys_platform == "darwin"
tensorflow-estimator
termcolor
terminado
testpath
tornado
traitlets
urllib3
wcwidth
webencodings
werkzeug
widgetsnbextension
wrapt
zipp
tflite-runtime==2.13.0     ; platform_system == "Linux"
onnxruntime>=1.18,<2
pytest==8.2.0
ruff==0.4.5
scikit-learn==1.5.0
pandas==2.2.2